+++ xmi2yang tool-v1.2-20160205/model/ObjectClass.js
