+++ xmi2yang tool-v1.2-20160205/model/OpenModelObject.js
@@ -1,0 +1,29 @@
+/********************************************************************************************************
+ * Name: UML to YANG Mapping Tool
+ * Copyright 2015 CAICT (China Academy of Information and Communication Technology (former China Academy of Telecommunication Research)). All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ *
+ * This tool is developed according to the mapping rules defined in onf2015.261_Mapping_Gdls_UML-YANG.04 by OpenNetworkFoundation(ONF) IMP group.
+ *
+ * file: \model\OpenModelObject.js
+ *
+ * The above copyright information should be included in all distribution, reproduction or derivative works of this software.
+ *
+ ****************************************************************************************************/
+function OpenModelObject(id,type,vr,cond,sup,inv,avcNot,dNot,cNot,passBR,opex,opid,ato){
+    this.id=id;
+    this.type=type;
+    this.valueRange=vr;
+    this.condition=cond;
+    this.status;
+    this.support=sup;
+    this.isInvariant=inv;
+    this.attributeValueChangeNotification=avcNot;
+    this.objectDeletionNotification=dNot;
+    this.objectCreationNotification=cNot;
+    this.passedByReference=passBR;
+    this["operation exceptions"]=opex;
+    this["isOperationIdempotent"]=opid;
+    this["isAtomic"]=ato;
+}
+module.exports=OpenModelObject;
